refactor: move OpenAI file helpers out of the assistants app
create_files_remote, _push_file_to_openai, _openai_create_file_with_retries
and get_and_store_openai_file are ordinary Files API calls with nothing
Assistants-specific about them. Their live consumer is
OpenAIRemoteIndexManager (remote vector-store indexes for Collections),
which outlives the Assistants feature, so they move to a new
llm_service/openai_files.py ahead of that app's removal.

chunk_list moves to the byte-identical apps.utils.deletion copy, and the
in-function import in upload_file_to_remote becomes module-level: the
cycle it guarded (assistants.sync -> index_managers) no longer exists.
That changes where the test must patch create_files_remote — at its use
site in index_managers, not at its definition.

Part of #4254 (Phase 1, task 1).
