+++ apps/assistants/sync.py
@@ -59,7 +59,6 @@ import contextlib
 import logging
 import pathlib
 from functools import wraps
-from io import BytesIO
 
 import openai
 from django.db.models import Count, Exists, OuterRef, Subquery
@@ -67,7 +66,6 @@ from django.forms import ValidationError
 from langchain_core.utils.function_calling import convert_to_openai_tool as lc_convert_to_openai_tool
 from openai import OpenAI
 from openai.types.beta import Assistant
-from tenacity import before_sleep_log, retry, retry_if_exception_type, stop_after_attempt, wait_exponential
 
 from apps.assistants.models import OpenAiAssistant, ToolResources
 from apps.assistants.utils import get_assistant_tool_options
@@ -75,6 +73,7 @@ from apps.documents.models import CollectionFile
 from apps.files.models import File, FilePurpose
 from apps.service_providers.exceptions import UnableToLinkFileException
 from apps.service_providers.llm_service.index_managers import OpenAIRemoteIndexManager
+from apps.service_providers.llm_service.openai_files import create_files_remote
 from apps.service_providers.models import LlmProvider, LlmProviderModel, LlmProviderTypes
 from apps.teams.models import Team
 from apps.utils.deletion import get_related_m2m_objects
@@ -547,52 +546,3 @@ def _openai_assistant_to_ocs_kwargs(assistant: Assistant, team=None, llm_provide
     if llm_provider:
         kwargs["llm_provider"] = llm_provider
     return kwargs
-
-
-def create_files_remote(client, files):
-    file_ids = []
-    for file in files:
-        if not file.external_id:
-            _push_file_to_openai(client, file)
-        file_ids.append(file.external_id)
-    return file_ids
-
-
-def _push_file_to_openai(client, file: File):
-    with file.file.open("rb") as fh:
-        bytesio = BytesIO(fh.read())
-    openai_file = _openai_create_file_with_retries(client, file.name, bytesio)
-    file.external_id = openai_file.id
-    file.external_source = "openai"
-    file.save()
-
-
-@retry(
-    wait=wait_exponential(multiplier=1, min=1, max=5),
-    reraise=True,
-    retry=retry_if_exception_type(openai.RateLimitError),
-    stop=stop_after_attempt(3),
-    before_sleep=before_sleep_log(logger, logging.INFO),
-)
-def _openai_create_file_with_retries(client, filename, bytesio):
-    logger.debug("Creating file in OpenAI: %s", filename)
-    return client.files.create(file=(filename, bytesio), purpose="assistants")
-
-
-def get_and_store_openai_file(client, file_id: str, team_id: int) -> File:
-    """Retrieve the content of the openai file with id=`file_id` and create a new `File` instance.
-
-    This is used at runtime to pull down files the assistant generates during a run
-    (code-interpreter outputs, generated images), which are attached to the chat as
-    conversation media — hence MESSAGE_MEDIA rather than ASSISTANT (bot config).
-    """
-    file = client.files.retrieve(file_id)
-    filename = file.filename
-    with contextlib.suppress(Exception):
-        filename = pathlib.Path(file.filename).name
-
-    file_content_obj = client.files.content(file_id)
-
-    return File.from_external_source(
-        filename, file_content_obj, file_id, "openai", team_id, purpose=FilePurpose.MESSAGE_MEDIA
-    )
